Move parsing DocxShapeLinePath inside class
