+++ Changelog.md
@@ -14,6 +14,7 @@
 ### Refactor
 * Move parsing Columns inside class
 * Move parsing TableGrid inside class
+* Move parsing DocxShapeLinePath inside class
 
 ## 0.1.2 (2016-06-07)
