--- lib/ooxml_parser/common_parser/common_data/alternate_content/drawing/graphic/shape/shape_properties/custom_geometry/docx_custom_geometry/docx_shape_line_element.rb
@@ -1,10 +1,0 @@
-# Docx Shape Line Element
-module OoxmlParser
-  class DocxShapeLineElement
-    attr_accessor :type, :points
-
-    def initialize(points = [])
-      @points = points
-    end
-  end
-end
