+++ lib/ooxml_parser/common_parser/common_data/alternate_content/drawing/graphic/shape/shape_properties/custom_geometry/ooxml_custom_geometry.rb
@@ -1,4 +1,3 @@
-require_relative 'docx_custom_geometry/docx_shape_line_element'
 require_relative 'docx_custom_geometry/docx_shape_line_path'
 # Docx Custom Geometry
 module OoxmlParser
